@@ -1,7 +1,5 @@
-/**
- * @type {import('@remix-run/dev').AppConfig}
- */
-module.exports = {
+/** @type {import('@remix-run/dev').AppConfig} */
+export default {
   cacheDirectory: './node_modules/.cache/remix',
-  ignoredRouteFiles: ['**/.*', '**/*.css', '**/*.test.{js,jsx,ts,tsx}'],
-}
+  ignoredRouteFiles: ['**/.*'],
+};
